pyproject.toml
[tool.poetry]
name = "multiobjective-opt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "multiobjective_opt"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.2"
jax = "^0.5.0"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
matplotlib = "^3.10.0"
ipython = "^8.31.0"
optax = "^0.2.4"
tqdm = "^4.67.1"
torch = "^2.5.1"
torchvision = "^0.20.1"
tabulate = "^0.9.0"
joblib = "^1.4.2"
pre-commit = "^4.1.0"
nsopy = "^1.52"
pytorch-ignite = "^0.5.1"
hydra-core = "^1.3.2"
mlflow = "^2.20.2"
fire = "^0.7.0"



[tool.black]
line-length = 90
target-version = ["py311"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
max-line-length = 90
ignore = ['E203', 'E501', 'W503', 'B950', 'B905']
exclude = ["*.ipynb"]
max-complexity = 12
select = ['B', 'C', 'E', 'F', 'W', 'B9']
per-file-ignores = [
    # for easier imports to __init__ without __all__
    '**/__init__.py: F401',
    # file to define custom types
    'types.py: F401',
]
count = true


[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"

[tool.nbqa.addopts]
flake8 = [
    "--extend-ignore=E203",
]

[tool.nbqa.exclude]
flake8 = "^notebooks/"

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
